requirements.txt
pandas
numpy
matplotlib
openpyxl
customtkinter
python-dateutil
geoip2
requests
tqdm
scapy
